# server/src/main/resources/swagger/api.0.1.yaml
openapi: 3.0.0
info:
  title: OpenAPI memorizer
  version: "0.1"
servers:
- url: http://localhost:8080
...
paths:
...
  /cards:
    post:
      description: create new card
      operationId: addNewCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDTO'
          description: card has added successfully
        "400":
          description: invalid card request passed
        "404":
          description: wrong deckId passed
      tags:
      - cards
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: cards
  /cards/{id}:
    delete:
      description: delete card by id
      operationId: deleteCardById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          example: 1234567890
          format: int32
          nullable: false
          type: integer
        style: simple
      responses:
        "200":
          description: card has deleted successfully
        "404":
          description: card with this id doesn't exist
      tags:
      - cards
      x-accepts: application/json
      x-tags:
      - tag: cards
    get:
      description: get card by id
      operationId: getCardById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          example: 1234567890
          format: int32
          nullable: false
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDTO'
          description: card has returned successfully
        "404":
          description: card with this id doesn't exist
      tags:
      - cards
      x-accepts: application/json
      x-tags:
      - tag: cards
    put:
      description: update card by id
      operationId: updateCardById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          example: 1234567890
          format: int32
          nullable: false
          type: integer
        style: simple
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
      responses:
        "200":
          description: card has successfully updated
        "404":
          description: unknown card id
        "400":
          description: invalid card request data passed
      tags:
      - cards
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: cards
components:
  schemas:
...
    SideDTO:
      example:
        imageLink: imageLink
        text: text
        title: title
      nullable: false
      x-class-extra-annotation: "@lombok.AllArgsConstructor"
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 100
          nullable: false
        text:
          type: string
          maxLength: 320
          nullable: true
        imageLink:
          type: string
          maxLength: 60
          nullable: true
      type: object
    CardDTO:
      example:
        question:
          imageLink: imageLink
          text: text
          title: title
        answer:
          imageLink: imageLink
          text: text
          title: title
        deckId: 1234567890
        id: 1234567890
      required:
        - id
        - deckId
        - question
        - answer
      properties:
        id:
          example: 1234567890
          format: int32
          nullable: false
          type: integer
        deckId:
          example: 1234567890
          format: int32
          nullable: false
          type: integer
        question:
          $ref: '#/components/schemas/SideDTO'
        answer:
          $ref: '#/components/schemas/SideDTO'
      type: object
    CardRequest:
      example:
        question:
          imageLink: imageLink
          text: text
          title: title
        answer:
          imageLink: imageLink
          text: text
          title: title
        deckId: 1234567890
      required:
        - deckId
        - question
        - answer
      properties:
        deckId:
          example: 1234567890
          format: int32
          nullable: false
          type: integer
        question:
          $ref: '#/components/schemas/SideDTO'
        answer:
          $ref: '#/components/schemas/SideDTO'
      type: object